POC move middleware to lib
