+++ packages/example-next-13/messages/en.json
@@ -5,8 +5,5 @@
   },
   "LocaleSwitcher": {
     "switchLocale": "Switch to {locale, select, de {German} en {English} other {Unknown}}"
-  },
-  "PageLayout": {
-    "pageTitle": "next-intl"
   }
 }
