+++ packages/example-next-13/src/components/LocaleSwitcher.tsx
@@ -1,11 +1,10 @@
 import {useLocale, useTranslations} from 'next-intl';
 import Link from 'next/link';
-import i18n from '../i18n';
 
 export default function LocaleSwitcher() {
   const t = useTranslations('LocaleSwitcher');
   const locale = useLocale();
-  const otherLocale = i18n.locales.find((cur) => cur !== locale);
+  const otherLocale = locale === 'en' ? 'de' : 'en';
 
   return (
     <Link href={'/' + otherLocale} prefetch={false}>
