+++ packages/example-next-13/src/components/PageLayout.tsx
@@ -1,5 +1,3 @@
-import {useTranslations} from 'next-intl';
-import Head from 'next/head';
 import {ReactNode} from 'react';
 
 type Props = {
@@ -8,13 +6,8 @@ type Props = {
 };
 
 export default function PageLayout({children, title}: Props) {
-  const t = useTranslations('PageLayout');
-
   return (
     <>
-      <Head>
-        <title>{[title, t('pageTitle')].join(' - ')}</title>
-      </Head>
       <div
         style={{
           padding: 24,
