--- packages/example-next-13/src/i18n.tsx
@@ -1,8 +1,0 @@
-import NextI18nConfig from './next-intl/I18nConfig';
-
-const i18n: NextI18nConfig = {
-  locales: ['en', 'de'],
-  defaultLocale: 'en'
-};
-
-export default i18n;
