+++ packages/example-next-13/src/middleware.tsx
@@ -1,4 +1,6 @@
-import i18n from './i18n';
-import createIntlMiddleware from './next-intl/server/createIntlMiddleware';
+import {createIntlMiddleware} from 'next-intl/server';
 
-export default createIntlMiddleware(i18n);
+export default createIntlMiddleware({
+  locales: ['en', 'de'],
+  defaultLocale: 'en'
+});
