--- packages/example-next-13/src/next-intl/README.md
@@ -1,1 +1,0 @@
-This folder contains a draft of APIs that will be moved to `next-intl`.
