+++ packages/next-intl/package.json
@@ -23,7 +23,8 @@
   "typings": "dist/src/index.d.ts",
   "files": [
     "dist",
-    "src"
+    "src",
+    "./server.js"
   ],
   "keywords": [
     "react",
@@ -42,6 +43,7 @@
     "testEnvironment": "jsdom"
   },
   "dependencies": {
+    "accept-language-parser": "^1.5.0",
     "use-intl": "^2.9.1"
   },
   "peerDependencies": {
@@ -54,7 +56,7 @@
     "dts-cli": "1.4.0",
     "eslint": "8.12.0",
     "eslint-config-molindo": "6.0.0",
-    "next": "^13.0.0",
+    "next": "^13.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "^4.4.4"
