+++ packages/next-intl/server.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/src/server';
