+++ packages/next-intl/src/server/NextI18nConfig.tsx
