+++ packages/next-intl/src/server/createIntlMiddleware.tsx
@@ -1,7 +1,10 @@
 import {NextRequest, NextResponse} from 'next/server';
-import NextI18nConfig from '../I18nConfig';
+import NextI18nConfig from './NextI18nConfig';
 import NextIntlCookie from './NextIntlCookie';
 import resolveLocale from './resolveLocale';
+
+// If there's an exact match for this path, we'll add the locale to the URL
+const REDIRECT_URL = '/';
 
 export default function createIntlMiddleware(i18n: NextI18nConfig) {
   return function middleware(request: NextRequest) {
@@ -15,8 +18,10 @@ export default function createIntlMiddleware(i18n: NextI18nConfig) {
     );
 
     let response;
-    if (request.nextUrl.pathname === '/') {
-      response = NextResponse.redirect(new URL('/' + locale, request.url));
+    if (request.nextUrl.pathname === REDIRECT_URL) {
+      response = NextResponse.redirect(
+        new URL(REDIRECT_URL + locale, request.url)
+      );
     } else {
       response = NextResponse.next();
     }
