+++ packages/next-intl/src/server/index.tsx
@@ -1,0 +1,1 @@
+export {default as createIntlMiddleware} from './createIntlMiddleware';
